Cleanup from review
