+++ libfuse/fs.go
@@ -201,22 +201,6 @@ func (f *FS) Statfs(ctx context.Context, req *fuse.StatfsRequest, resp *fuse.Sta
 	return nil
 }
 
-func (f *FS) writeAccessError(ctx context.Context, h *libkbfs.TlfHandle, filename string) error {
-	username, _, err := f.config.KBPKI().GetCurrentUserInfo(ctx)
-	if err != nil {
-		return err
-	}
-	return libkbfs.NewWriteAccessError(h, username, filename)
-}
-
-func (f *FS) writeUnsupportedError(ctx context.Context, filename string) error {
-	username, _, err := f.config.KBPKI().GetCurrentUserInfo(ctx)
-	if err != nil {
-		return err
-	}
-	return libkbfs.NewWriteUnsupportedError(username, filename)
-}
-
 // Root represents the root of the KBFS file system.
 type Root struct {
 	private *FolderList
@@ -279,14 +263,14 @@ var _ fs.NodeCreater = (*Root)(nil)
 func (r *Root) Create(ctx context.Context, req *fuse.CreateRequest, resp *fuse.CreateResponse) (_ fs.Node, _ fs.Handle, err error) {
 	r.log().CDebugf(ctx, "FS Create")
 	defer func() { r.private.fs.reportErr(ctx, libkbfs.WriteMode, err) }()
-	return nil, nil, r.private.fs.writeUnsupportedError(ctx, libkbfs.BuildCanonicalPath(r.PathType(), req.Name))
+	return nil, nil, libkbfs.NewWriteUnsupportedError(libkbfs.BuildCanonicalPath(r.PathType(), req.Name))
 }
 
 // Mkdir implements the fs.NodeMkdirer interface for Root.
 func (r *Root) Mkdir(ctx context.Context, req *fuse.MkdirRequest) (_ fs.Node, err error) {
 	r.log().CDebugf(ctx, "FS Mkdir")
 	defer func() { r.private.fs.reportErr(ctx, libkbfs.WriteMode, err) }()
-	return nil, r.private.fs.writeUnsupportedError(ctx, libkbfs.BuildCanonicalPath(r.PathType(), req.Name))
+	return nil, libkbfs.NewWriteUnsupportedError(libkbfs.BuildCanonicalPath(r.PathType(), req.Name))
 }
 
 var _ fs.Handle = (*Root)(nil)
