+++ libkbfs/errors.go
@@ -162,7 +162,7 @@ type ReadAccessError struct {
 // Error implements the error interface for ReadAccessError
 func (e ReadAccessError) Error() string {
 	return fmt.Sprintf("%s does not have read access to directory %s",
-		e.User, buildCanonicalPath(e.Public, e.Tlf))
+		e.User, buildCanonicalPathForTlfName(e.Public, e.Tlf))
 }
 
 // WriteAccessError indicates an error when trying to write a file
@@ -177,20 +177,19 @@ type WriteAccessError struct {
 func (e WriteAccessError) Error() string {
 	if e.Tlf != "" {
 		return fmt.Sprintf("%s does not have write access to directory %s",
-			e.User, buildCanonicalPath(e.Public, e.Tlf))
+			e.User, buildCanonicalPathForTlfName(e.Public, e.Tlf))
 	}
 	return fmt.Sprintf("%s does not have write access to %s", e.User, e.Filename)
 }
 
 // WriteUnsupportedError indicates an error when trying to write a file
 type WriteUnsupportedError struct {
-	User     libkb.NormalizedUsername
 	Filename string
 }
 
 // Error implements the error interface for WriteAccessError
 func (e WriteUnsupportedError) Error() string {
-	return fmt.Sprintf("%s can't write to %s", e.User, e.Filename)
+	return fmt.Sprintf("Writing to %s is unsupported", e.Filename)
 }
 
 // NewReadAccessError constructs a ReadAccessError for the given
@@ -217,9 +216,8 @@ func NewWriteAccessError(h *TlfHandle, username libkb.NormalizedUsername, filena
 }
 
 // NewWriteUnsupportedError returns unsupported error trying to write a file
-func NewWriteUnsupportedError(username libkb.NormalizedUsername, filename string) error {
+func NewWriteUnsupportedError(filename string) error {
 	return WriteUnsupportedError{
-		User:     username,
 		Filename: filename,
 	}
 }
@@ -235,7 +233,7 @@ type NeedSelfRekeyError struct {
 func (e NeedSelfRekeyError) Error() string {
 	return fmt.Sprintf("This device does not yet have read access to "+
 		"directory %s, log into Keybase from one of your other "+
-		"devices to grant access", buildCanonicalPath(false, e.Tlf))
+		"devices to grant access", buildCanonicalPathForTlfName(false, e.Tlf))
 }
 
 // NeedOtherRekeyError indicates that the folder in question needs to
@@ -250,7 +248,7 @@ func (e NeedOtherRekeyError) Error() string {
 	return fmt.Sprintf("This device does not yet have read access to "+
 		"directory %s, ask one of the other directory participants to "+
 		"log into Keybase to grant you access automatically",
-		buildCanonicalPath(false, e.Tlf))
+		buildCanonicalPathForTlfName(false, e.Tlf))
 }
 
 // NotFileBlockError indicates that a file block was expected but a
